CSP-2118-2119 - Addressed PR comments

+++ src/SFA.DAS.DownloadService.Api.Types/Roatp/RoatpResult.cs
@@ -5,7 +5,7 @@ namespace SFA.DAS.DownloadService.Api.Types.Roatp
 {
     public class RoatpResult
     {
-        public long Ukprn { get; set; }
+        public int Ukprn { get; set; }
 
         [JsonProperty("Organisation Name")]
         public string OrganisationName { get; set; }
